+++ src/AbstractFileGenerator.php
@@ -6,27 +6,30 @@ namespace MetaRush\Getter;
 
 abstract class AbstractFileGenerator
 {
+    /**
+     *
+     * @var FileGeneratorConfig
+     */
+    protected $cfg;
     private $syntaxGenerator;
 
-    public function __construct(SyntaxGenerator $syntaxGenerator)
+    public function __construct(FileGeneratorConfig $cfg, SyntaxGenerator $syntaxGenerator)
     {
+        $this->cfg = $cfg;
         $this->syntaxGenerator = $syntaxGenerator;
     }
 
     /**
      * Generate class file
      *
-     * @param string $className name of class to generate
-     * @param array $data PHP data to generated from
-     * @param string $location where to save generated class file
      * @return void
      */
-    protected function generateClassFile(string $className, array $data, string $location): void
+    protected function generateClassFile(): void
     {
         $header = "<?php\n\ndeclare(strict_types=1);\n\n";
 
-        $classSyntax = $this->syntaxGenerator->classSyntax($className, $data);
+        $classSyntax = $this->syntaxGenerator->classSyntax($this->cfg->getClassName(), $this->cfg->getData(), $this->cfg->getExtendedClass());
 
-        \file_put_contents($location . $className . '.php', $header . $classSyntax);
+        \file_put_contents($this->cfg->getLocation() . $this->cfg->getClassName() . '.php', $header . $classSyntax);
     }
 }