+++ tests/unit/YamlAdapterTest.php
@@ -6,13 +6,16 @@ use PHPUnit\Framework\TestCase;
 
 class YamlAdapterTest extends TestCase
 {
+    private $cfg;
     private $fileGenerator;
 
     public function setUp(): void
     {
+        $this->cfg = new \MetaRush\Getter\FileGeneratorConfig;
+
         $syntaxGenerator = new \MetaRush\Getter\SyntaxGenerator;
 
-        $this->fileGenerator = new \MetaRush\Getter\Adapters\Yaml($syntaxGenerator);
+        $this->fileGenerator = new \MetaRush\Getter\Adapters\Yaml($this->cfg, $syntaxGenerator);
     }
 
     public function testGenerateClassFile()
@@ -20,7 +23,10 @@ class YamlAdapterTest extends TestCase
         $location = __DIR__ . '/samples/';
         $yamlFile = $location . 'sample.yaml';
 
-        $this->fileGenerator->generate('FooYaml', $yamlFile, $location);
+        $this->cfg->setClassName('FooYaml');
+        $this->cfg->setLocation($location);
+
+        $this->fileGenerator->generate($yamlFile);
 
         $this->assertFileExists($location . 'FooYaml.php');