+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # CHANGELOG
 
+## 1.1.0 - 2019-06-06
+
+### Added
+
+- Add setExtendedClass() config method.
+
 ## 1.0.1 - 2019-06-03
 
 ### Fixed
 
-- Move phpunit to require-dev only
+- Move phpunit to require-dev only.
 
 ## 1.0.0 - 2019-06-03